Add lock-files script
Add a script to create and update the minimal and recent lock files. On
the way create a `stdlib.sh` script to reduce code duplication and also
a `template.sh` to make writing the next script easier.

Note I could not get `flag_verbose` to work with sourcing the `stdlib`
script without exporting a global variable which I didn't do so as not
to pollute the environment.

If this merges we can remove the `update-lock-files.sh` script from
other repos and use this new script.