Add -j1 to docs build invocation
After upgrade of `hashes` but before we upgrade it in `secp` we get
multiple versions in the dep tree. This makes the docs build choke
because output directory uses the name of the crate (without the
version) which then conflicts and causes the build to fail.

Adding `-j1` so that only a single job is run when building the docs
resolves the issue.