Add support for corepc
Currently the `corpec` repository excludes the `node` crate because it
has unusual feature requirements - it requires at least one version
feature.

Add support for testing `node`. The logic in `build_and_test` is general
and uses a new var defined in `contrib/test_vars.sh` but does not
require this var to exist.

The logic in `do_lint` is ugly and specific to `node`. I couldn't think
of a cleaner way to do it.