ci: Build and test with all features
We test various combinations of features in the feature matrix but
somehow we forgot to build and test with `--all-features`. We do lint
with all features so this is not a massive fail but its a fail none the
less.

Discovered by Andrew's CI while reviewing recent changes to `corepc`.